rename client module and introduce publisher module
