+++ src/collectors.rs
@@ -1,4 +1,4 @@
-pub mod collector_exceptions;
+pub mod collector_exceptions; // TODO make this non-public
 pub mod dmi_collector;
 pub mod network_collector;
-pub mod util;
+mod util; // TODO make this non-public
