+++ src/publisher.rs
@@ -1,2 +1,3 @@
+pub mod api_client; // TODO make this non-public
 pub mod publisher;
-pub mod publisher_exceptions;
+pub mod publisher_exceptions; // TODO make this non-public
